fix overflow issue in x position calculation

+++ gym_super_mario_bros/smb_env.py
@@ -147,7 +147,7 @@ class SuperMarioBrosEnv(NESEnv):
         # TODO: resolve RuntimeWarning: overflow encountered in ubyte_scalars
         # subtract the left x position 0x071c from the current x 0x86
         # return (self.ram[0x86] - self.ram[0x071c]) % 256
-        return np.uint8(int(self.ram[0x86]) - int(self.ram[0x071c])) % 256
+        return (int(self.ram[0x86]) - int(self.ram[0x071c])) % 256
 
     @property
     def _y_pixel(self):
